+++ lib/src/widgets/raw_slide_countdown.dart
@@ -1,13 +1,20 @@
 import 'package:flutter/material.dart';
 import 'package:slide_countdown/slide_countdown.dart';
 
+typedef RawSlideCountdownBuilder = Widget Function(
+  BuildContext context,
+  Duration duration,
+);
+
 class RawSlideCountdown extends StatefulWidget {
   const RawSlideCountdown({
     super.key,
     required this.streamDuration,
+    required this.builder,
   });
 
   final StreamDuration streamDuration;
+  final RawSlideCountdownBuilder builder;
 
   @override
   State<RawSlideCountdown> createState() => _RawSlideCountdownState();
@@ -20,9 +27,7 @@ class _RawSlideCountdownState extends State<RawSlideCountdown> {
       stream: widget.streamDuration.durationLeft,
       builder: (_, snapshoot) {
         if (snapshoot.hasData) {
-          return Wrap(
-            children: [],
-          );
+          return widget.builder(context, snapshoot.data!);
         }
 
         return SizedBox.shrink();